dispatcher: generic app event dispatching (#346)
